# static/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "VoiceVerse API",
    "description": "AI-powered text-to-speech API for converting text into natural, high-quality audio. Supports multiple voices, file management, and advanced features like smart chunking and preprocessing.",
    "version": "1.0.0",
    "contact": {
      "name": "VoiceVerse Support",
      "url": "https://voiceverse.app"
    },
    "license": {
      "name": "Proprietary"
    }
  },
  "servers": [
    {
      "url": "https://voiceverse.app",
      "description": "Production server"
    },
    {
      "url": "https://127.0.0.1:5000",
      "description": "Local development server"
    }
  ],
  "paths": {
...
    "/api/agent/preprocess": {
      "post": {
        "summary": "AI text preprocessing",
        "description": "Use AI to clean and optimize text for better speech synthesis",
        "operationId": "preprocessText",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Text preprocessed successfully"
          }
        },
        "tags": ["AI Agents"]
      }
    },
    "/api/agent/suggest-metadata": {
      "post": {
        "summary": "AI metadata suggestions",
        "description": "Get AI-generated suggestions for filename, category, and voice based on text content",
        "operationId": "suggestMetadata",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Metadata suggestions generated"
          }
        },
        "tags": ["AI Agents"]
      }
    },
    "/api/agent/analyze": {
      "post": {
        "summary": "AI text analysis",
        "description": "Analyze text for quality issues and get recommendations for better speech synthesis",
        "operationId": "analyzeText",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "properties": {
                  "text": {"type": "string"}
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Text analysis completed"
          }
        },
        "tags": ["AI Agents"]
      }
    },
...
  },
...
}